Title: 세 수의 합과 평균 구하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240119/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.cpp
@@ -9,6 +9,6 @@ int main() {
     cin >> c;
     cout << fixed;
     cout.precision(0);
-    cout << a+b+c << endl << (double)(a+b+c)/3;
+    cout << a+b+c << endl << (int)(a+b+c)/3;
     return 0;
 }
